+++ contracts/src/token/test/utils/address.ts
@@ -1,5 +1,5 @@
 import {
-  convert_bigint_to_Uint8Array,
+  convertFieldToBytes,
   encodeCoinPublicKey,
 } from '@midnight-ntwrk/compact-runtime';
 import { encodeContractAddress } from '@midnight-ntwrk/ledger';
@@ -62,7 +62,7 @@ export const createEitherTestContractAddress = (str: string) => ({
 });
 
 export const zeroUint8Array = (length = 32) =>
-  convert_bigint_to_Uint8Array(length, 0n);
+  convertFieldToBytes(length, 0n, "");
 
 export const ZERO_KEY = {
   is_left: true,